requirements.txt
streamlit 
streamlit-autorefresh
psycopg2-binary
pandas
plotly
kafka-python
faker
python-dotenv